+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+- `FnoFunctionModelProvider`: Implementation locations can be changed w.r.t. what's in the function descriptions by
+  providing a map `old location -> new location`. 
+  This feature allows to use the right location at runtime without changing the function descriptions.
+
 ## [0.0.4] - 2022-05-17
 
 ### Fixed